+++ src/popup/pages/Networks/NetworkSelect/index.tsx
@@ -63,6 +63,8 @@ const NetworkSelect = () => {
           dispatch(updateTokenAssets([]));
           dispatch(updateStakingRefresh(true));
           dispatch(updateShouldRequest(true));
+        } else {
+          dispatch(updateShouldRequest(true, true));
         }
 
         sendNetworkChangeMsg(newConfig as unknown as Parameters<typeof sendNetworkChangeMsg>[0]);