+++ src/popup/pages/Send/tokenSign/index.tsx
@@ -77,10 +77,6 @@ const TokenSignPage = () => {
 
   const { fetchAccountData } = useFetchAccountData(currentAccount as Parameters<typeof useFetchAccountData>[0]);
 
-  const currentAddress = useAppSelector(
-    (state) => state.accountInfo.currentAccount.address
-  );
-
   const fetchAccountInfo = useCallback(async () => {
     if (isShowLoading.current) {
       Loading.show();
@@ -89,7 +85,7 @@ const TokenSignPage = () => {
     await fetchAccountData();
     isFirstRequest.current = false;
     Loading.hide();
-  }, [dispatch, currentAddress]);
+  }, [dispatch, fetchAccountData]);
 
   const getSignParams = useCallback(() => {
     sendMsg(